+++ Profile/func/Mini-Functions.ps1
@@ -18,7 +18,7 @@ function Mini-Functions {
         .NOTES
             Author: Kai Krutscho
         .LINK
-            https://www.github.com/Kaimodo/PowerShell-Profile
+            [1]: 	https://www.github.com/Kaimodo/PowerShell-Profile
         #>
         [CmdletBinding()]
         Param()